% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_AFN}
\alias{ROM_Coul_AFN}
\title{Couleur de la region \if{html}{\out{<span style="color:rgb(233, 78, 27);">}}Afrique du Nord\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_AFN
}
\description{
Couleur de la region \if{html}{\out{<span style="color:rgb(233, 78, 27);">}}Afrique du Nord\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
