% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_AMC}
\alias{ROM_Coul_AMC}
\title{Couleur de la region \if{html}{\out{<span style="color:rgb(134, 196, 236);">}}Amerique Centrale\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_AMC
}
\description{
Couleur de la region \if{html}{\out{<span style="color:rgb(134, 196, 236);">}}Amerique Centrale\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
