% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_ASN}
\alias{ROM_Coul_ASN}
\title{Couleur de la region \if{html}{\out{<span style="color:rgb(180, 151, 199);">}}Asie du Nord Est\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_ASN
}
\description{
Couleur de la region \if{html}{\out{<span style="color:rgb(180, 151, 199);">}}Asie du Nord Est\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
