% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_CEI}
\alias{ROM_Coul_CEI}
\title{Couleur de la region de \if{html}{\out{<span style="color:rgb(173, 204, 83);">}}CEI (CIS en anglais)\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_CEI
}
\description{
Couleur de la region de \if{html}{\out{<span style="color:rgb(173, 204, 83);">}}CEI (CIS en anglais)\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
