% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_CHN}
\alias{ROM_Coul_CHN}
\title{Couleur de la \if{html}{\out{<span style="color:rgb(158, 117, 178);">}}Chine\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_CHN
}
\description{
Couleur de la \if{html}{\out{<span style="color:rgb(158, 117, 178);">}}Chine\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
