% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_EURC}
\alias{ROM_Coul_EURC}
\title{Couleur des regions \if{html}{\out{<span style="color:rgb(149, 193, 31);">}}Europe et de la CEI\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_EURC
}
\description{
Couleur des regions \if{html}{\out{<span style="color:rgb(149, 193, 31);">}}Europe et de la CEI\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
