% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_RT}
\alias{ROM_Coul_RT}
\title{Couleur des \if{html}{\out{<span style="color:rgb(255, 152, 0);">}}Turboreacteurs\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_RT
}
\description{
Couleur des \if{html}{\out{<span style="color:rgb(255, 152, 0);">}}Turboreacteurs\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
