% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couleurs.R
\docType{data}
\name{ROM_Coul_WB}
\alias{ROM_Coul_WB}
\title{Couleur des \if{html}{\out{<span style="color:rgb(112, 48, 160);">}}Wide Body\if{html}{\out{</span>}} pour le LTMF}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROM_Coul_WB
}
\description{
Couleur des \if{html}{\out{<span style="color:rgb(112, 48, 160);">}}Wide Body\if{html}{\out{</span>}} pour le LTMF
}
\keyword{datasets}
